// web/src/lib/api/schemas/scheduleEndpoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flightpath API
 * Live aircraft positions from the OpenSky Network with flight, airport and airline metadata.
 * OpenAPI spec version: dev
 */

export interface ScheduleEndpoint {
  actual?: string
  airport?: string
  /** Delay in minutes. */
  delayMin?: number
  estimated?: string
  gate?: string
  iata?: string
  icao?: string
  scheduled?: string
  terminal?: string
  timezone?: string
}